requirements.txt
# Core dependencies
ccxt==4.1.0
asyncio==3.4.3
aiohttp==3.9.0
websockets==12.0
websocket-client==1.7.0

# Data processing
pandas==2.0.3
numpy==1.24.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Redis for queues
redis==5.0.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.2

# Logging
colorlog==6.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Monitoring
prometheus-client==0.19.0

# Telegram alerts (optional)
python-telegram-bot==20.7
